requirements.txt
fast_bitrix24
fastapi[standard]
python-dotenv
dadata
sqlalchemy
pyodbc
pytest
requests
loguru
pydantic
pydantic-settings
git+https://github.com/vvpreo/fastapi-jwt-auth.git